moved mock data out of pages
